changed type

+++ chapter_09_to_11_user_input_and_conditional_statements/question_04/main.py
@@ -1,4 +1,4 @@
-remaining_fuel : int = int(input("Enter the remaining fuel in car (in litres): "))
+remaining_fuel : float = float(input("Enter the remaining fuel in car (in litres): "))
 
 if remaining_fuel < 0.25:
     print("Please refill the fuel in your car")
